requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mecab-python3==1.0.3
mysql-connector-python==8.0.23
numpy==1.20.2
packaging==20.9
pandas==1.2.3
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
uritemplate==3.0.1
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
unidic-lite